# docs/swagger.yaml
basePath: /
definitions:
  auth.BindRefresh:
    properties:
      refresh_token:
        type: string
    required:
    - refresh_token
    type: object
  auth.GetInfo:
    properties:
      name:
        type: string
      phone_number:
        type: string
      user_id:
        type: integer
    type: object
  auth.IfSuccessReset:
    properties:
      message:
        type: string
      status:
        $ref: '#/definitions/constants.Status'
    type: object
  auth.Needlogin:
    properties:
      password:
        type: string
      phone_number:
        type: string
    required:
    - password
    - phone_number
    type: object
  auth.RegisterIn:
    properties:
      name:
        type: string
      password:
        type: string
      phone_number:
        type: string
    required:
    - name
    - password
    - phone_number
    type: object
  auth.ResetModel:
    properties:
      new_password:
        type: string
      old_password:
        type: string
      phone_number:
        type: string
    required:
    - new_password
    - old_password
    - phone_number
    type: object
  cerror.CustomError400:
    properties:
      message:
        example: 입력하신 부분을 다시 확인해주세요
        type: string
      status_code:
        example: 400
        type: integer
    type: object
  cerror.CustomError401:
    properties:
      message:
        example: 인증에 실패했습니다 다시 로그인 해주세요
        type: string
      status_code:
        example: 401
        type: integer
    type: object
  cerror.CustomError500:
    properties:
      message:
        example: 예기치 않은 오류
        type: string
      status_code:
        example: 500
        type: integer
    type: object
  constants.Status:
    enum:
    - Ok
    - Fail
    type: string
    x-enum-varnames:
    - StatusOk
    - StatusFail
  middleware.AccessAndRefreshResponse:
    properties:
      access_token:
        type: string
      expires_at:
        type: string
      refresh_token:
        type: string
    type: object
  middleware.AccessTokenResponse:
    properties:
      access_token:
        type: string
      expires_at:
        type: string
    type: object
info:
  contact: {}
  description: This is a sample server to dooluck
  title: Swagger gin-login
  version: "1.0"
paths:
  /api/auth/info:
    get:
      consumes:
      - application/json
      description: 로그인한 자기 정보 가져오기
      parameters:
      - description: access token
        in: header
        name: auth-token
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/auth.GetInfo'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/cerror.CustomError400'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/cerror.CustomError401'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/cerror.CustomError500'
      summary: 로그인정보
      tags:
      - auth
  /api/auth/leave/{pwd}:
    delete:
      consumes:
      - application/json
      description: 회원 탈퇴
      parameters:
      - description: access token
        in: header
        name: auth-token
        required: true
        type: string
      - description: 패스워드
        in: path
        name: pwd
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/cerror.CustomError400'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/cerror.CustomError401'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/cerror.CustomError500'
      summary: 회원탈퇴
      tags:
      - auth
  /api/auth/login:
    post:
      consumes:
      - application/json
      parameters:
      - description: 전화번호, 비밀번호
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/auth.Needlogin'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/middleware.AccessAndRefreshResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/cerror.CustomError400'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/cerror.CustomError401'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/cerror.CustomError500'
      summary: login기능
      tags:
      - auth
  /api/auth/logout:
    post:
      consumes:
      - application/json
      parameters:
      - description: access token
        in: header
        name: auth-token
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/cerror.CustomError400'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/cerror.CustomError401'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/cerror.CustomError500'
      summary: logout 기능
      tags:
      - auth
  /api/auth/refresh-token:
    post:
      consumes:
      - application/json
      description: refresh token 으로 access token 갱신
      parameters:
      - description: access token
        in: header
        name: auth-token
        required: true
        type: string
      - description: 갱신
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/auth.BindRefresh'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/middleware.AccessTokenResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/cerror.CustomError400'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/cerror.CustomError401'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/cerror.CustomError500'
      summary: refresh token
      tags:
      - auth
  /api/auth/register:
    post:
      consumes:
      - application/json
      parameters:
      - description: 전화번호,비밀번호,이름
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/auth.RegisterIn'
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/cerror.CustomError400'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/cerror.CustomError401'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/cerror.CustomError500'
      summary: register
      tags:
      - auth
  /api/auth/reset-password:
    post:
      consumes:
      - application/json
      description: 비밀번호 초기화
      parameters:
      - description: 전화번호, 비밀번호
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/auth.ResetModel'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/auth.IfSuccessReset'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/cerror.CustomError400'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/cerror.CustomError401'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/cerror.CustomError500'
      summary: 패스워드 초기화
      tags:
      - auth
swagger: "2.0"